feat: add greetings & goodbying
